+++ LibraryBackend/src/main/java/com/library/book/service/BookService.java
@@ -9,7 +9,7 @@ public interface BookService {
 
 	List<BookDTO>BookSearch(String Search);
 	
-	List<BookDTO>BookDetail(int bookNumberInfo);
+	List<BookDTO>BookDetail(Long bookNumberInfo);
 
 	List<BookDTO>BookCategory();
